--- src/main/java/cucumberframework/PracticeProjectWithCucumber/App.java
@@ -1,13 +1,0 @@
-package cucumberframework.PracticeProjectWithCucumber;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
