--- src/test/java/ActionsofParticularPage/HomePageActions.java
@@ -1,7 +1,0 @@
-package ActionsofParticularPage;
-
-public class HomePageActions {
-	
-	//writing code and commit code using git bush
-
-}
