--- src/test/java/pages/Page2.java
@@ -1,5 +1,0 @@
-package pages;
-
-public class Page2 {
-
-}
